+++ features/saved-tabs/components/ProjectUrlItem.tsx
@@ -1,6 +1,17 @@
 import { useSortable } from '@dnd-kit/sortable'
 import { CSS } from '@dnd-kit/utilities'
 import { ChevronRight, GripVertical, X } from 'lucide-react'
+import { useState } from 'react'
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import type { CustomProject, UserSettings } from '@/types/storage'
@@ -48,6 +59,8 @@ export const ProjectUrlItem = ({
   // 実際のURLを保存（元のURL）
   const originalUrl = item.url
 
+  const [isDeleteConfirmOpen, setIsDeleteConfirmOpen] = useState(false)
+
   // ドラッグアンドドロップの設定を強化
   const {
     attributes,
@@ -89,84 +102,107 @@ export const ProjectUrlItem = ({
   const isInSubcategory = categoryLevel > 0
 
   return (
-    <li
-      ref={setNodeRef}
-      style={style}
-      className={`group relative flex min-w-0 items-center overflow-hidden border-border border-b pb-1 last:border-0 ${isDragging ? 'bg-secondary/50 opacity-50' : ''}
+    <>
+      <li
+        ref={setNodeRef}
+        style={style}
+        className={`group relative flex min-w-0 items-center overflow-hidden border-border border-b pb-1 last:border-0 ${isDragging ? 'bg-secondary/50 opacity-50' : ''}
         ${isInSubcategory ? 'pl-2' : ''}
         ${item.category ? 'border-l-2 border-l-primary/30' : ''}
       `}
-      data-url={originalUrl}
-      data-project-id={projectId}
-      data-category={item.category}
-      data-has-category={!!item.category}
-      data-category-level={categoryLevel}
-      data-parent-type={parentType || ''}
-      data-in-uncategorized={isInUncategorizedArea ? 'true' : 'false'}
-    >
-      <div
-        {...attributes}
-        {...listeners}
-        className='cursor-grab p-1 opacity-30 active:cursor-grabbing group-hover:opacity-100'
+        data-url={originalUrl}
+        data-project-id={projectId}
+        data-category={item.category}
+        data-has-category={!!item.category}
+        data-category-level={categoryLevel}
+        data-parent-type={parentType || ''}
+        data-in-uncategorized={isInUncategorizedArea ? 'true' : 'false'}
       >
-        <GripVertical size={16} className='text-muted-foreground' />
-      </div>
-      {/* タイトル＋バッジ部 */}
-      <div className='flex min-w-0 flex-1 items-center'>
-        <Button
-          asChild
-          variant='ghost'
-          size='sm'
-          className='flex flex-1 items-center gap-1 overflow-hidden text-left text-foreground hover:text-foreground hover:underline'
+        <div
+          {...attributes}
+          {...listeners}
+          className='cursor-grab p-1 opacity-30 active:cursor-grabbing group-hover:opacity-100'
         >
-          <a
-            href={item.url}
-            target='_blank'
-            rel='noopener noreferrer'
+          <GripVertical size={16} className='text-muted-foreground' />
+        </div>
+        {/* タイトル＋バッジ部 */}
+        <div className='flex min-w-0 flex-1 items-center'>
+          <Button
+            asChild
+            variant='ghost'
+            size='sm'
+            className='flex flex-1 items-center gap-1 overflow-hidden text-left text-foreground hover:text-foreground hover:underline'
+          >
+            <a
+              href={item.url}
+              target='_blank'
+              rel='noopener noreferrer'
+              onClick={e => {
+                e.preventDefault()
+                handleOpenUrl(item.url)
+              }}
+            >
+              {/* サブカテゴリ付きのURLの場合はChevronRightを表示 */}
+              {item.category?.includes('/') && (
+                <ChevronRight
+                  size={14}
+                  className='mr-1 inline-block text-primary'
+                />
+              )}
+              <span className='flex-1 truncate'>{item.title || item.url}</span>
+              {/* カテゴリ階層の視覚的な表示をシンプル化 */}
+              {item.category?.includes('/') && (
+                <Badge variant='outline' className='ml-2 shrink-0 text-xs'>
+                  {getCategoryDisplayName(item.category)}
+                </Badge>
+              )}
+            </a>
+          </Button>
+        </div>
+        {/* ボタン群 */}
+        <div className='flex shrink-0 items-center gap-1 opacity-0 transition-opacity duration-200 group-hover:opacity-100'>
+          <Button
+            variant='ghost'
+            size='sm'
             onClick={e => {
               e.preventDefault()
-              handleOpenUrl(item.url)
+              e.stopPropagation()
+              if (!settings.confirmDeleteEach) {
+                handleDeleteUrl(projectId, item.url)
+              } else {
+                setIsDeleteConfirmOpen(true)
+              }
             }}
+            className='h-8 w-8 cursor-pointer p-0'
+            title='URLを削除'
+            aria-label='URLを削除'
           >
-            {/* サブカテゴリ付きのURLの場合はChevronRightを表示 */}
-            {item.category?.includes('/') && (
-              <ChevronRight
-                size={14}
-                className='mr-1 inline-block text-primary'
-              />
-            )}
-            <span className='flex-1 truncate'>{item.title || item.url}</span>
-            {/* カテゴリ階層の視覚的な表示をシンプル化 */}
-            {item.category?.includes('/') && (
-              <Badge variant='outline' className='ml-2 shrink-0 text-xs'>
-                {getCategoryDisplayName(item.category)}
-              </Badge>
-            )}
-          </a>
-        </Button>
-      </div>
-      {/* ボタン群 */}
-      <div className='flex shrink-0 items-center gap-1 opacity-0 transition-opacity duration-200 group-hover:opacity-100'>
-        <Button
-          variant='ghost'
-          size='sm'
-          onClick={e => {
-            e.preventDefault()
-            e.stopPropagation()
-            if (
-              !settings.confirmDeleteEach ||
-              window.confirm('このURLを削除しますか？')
-            ) {
-              handleDeleteUrl(projectId, item.url)
-            }
-          }}
-          className='h-8 w-8 cursor-pointer p-0'
-          title='URLを削除'
-          aria-label='URLを削除'
-        >
-          <X size={14} />
-        </Button>
-      </div>
-    </li>
+            <X size={14} />
+          </Button>
+        </div>
+      </li>
+
+      <AlertDialog
+        open={isDeleteConfirmOpen}
+        onOpenChange={setIsDeleteConfirmOpen}
+      >
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>URLを削除しますか？</AlertDialogTitle>
+            <AlertDialogDescription>
+              このURLを削除します。この操作は元に戻せません。
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>キャンセル</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={() => handleDeleteUrl(projectId, item.url)}
+            >
+              削除する
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </>
   )
 }
